update: rearrange limit related code to single file
